requirements.txt
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
DateTime==5.4
docopt==0.6.2
flash==1.0.3
Flask==3.0.2
funcs==0.8.0
gitdb==4.0.11
GitPython==3.1.42
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
named==1.3.0
packaging==23.2
pipreqs==0.4.13
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
smmap==5.0.1
solus==1.1.0
typing-aliases==1.6.0
typing_extensions==4.9.0
urllib3==2.2.0
Werkzeug==3.0.1
wraps==0.8.0
yarg==0.1.9
zope.interface==6.2
